binary-clock/requirements.txt
cbor2==5.6.4
luma.core==2.4.2
luma.led-matrix==1.7.1
pillow==10.4.0
pyftdi==0.55.4
pyserial==3.5
pyusb==1.2.1
RPi.GPIO==0.7.1
smbus2==0.4.3
spidev==3.6
